LibGfx/JPEG2000: Support [8..16] bits per pixel
Do this by making decode_code_block() write to a float buffer,
instead of an i16 buffer.

Alternatively, we could make decode_code_block() stop after some
number of bits since we only need 8 anyways and update a few places
in JPEG2000.cpp that currently read bit_depth(), but this here seems
perf-neutral (even slightly positive) and is simpler, so let's go
with this for now.

(16 bpp images are very rare, and I haven't seen images with more than
16 bpp, even though the spec allows up to 38 bpp.)

+++ Userland/Libraries/LibGfx/ImageFormats/JPEG2000BitplaneDecoding.h
@@ -54,7 +54,7 @@ inline auto number_of_passes_from_segment_index_in_bypass_mode(unsigned segment_
     return segment_index % 2 == 1 ? 2u : 1u;
 }
 
-inline ErrorOr<void> decode_code_block(Span2D<i16> result, SubBand sub_band, int number_of_coding_passes, Vector<ReadonlyBytes, 1> segments, int M_b, int p, BitplaneDecodingOptions options = {})
+inline ErrorOr<void> decode_code_block(Span2D<float> result, SubBand sub_band, int number_of_coding_passes, Vector<ReadonlyBytes, 1> segments, int M_b, int p, BitplaneDecodingOptions options = {})
 {
     // This is an implementation of the bitplane decoding algorithm described in Annex D of the JPEG2000 spec.
     // It's modeled closely after Figure D.3 – Flow chart for all coding passes on a code-block bit-plane,
